[BUGFIX] Fix logout handling on TYPO3v12
In TYPO3v12, "logoff_post_processing" hook has been deprecated.
We migrate to the "BeforeUserLogoutEvent" which allows us to clean
up sensitive data from the session as before.

For admin users, the Install tool session handler is invoked during
the logout process. The session save path gets modified during that.
To be able to clean the session from sensitive data, we have to
store the original session save path in the backend user session.

Changelog:
https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.3/Deprecation-100307-VariousHooksRelatedToAuthenticationUsers.html
https://docs.typo3.org/c/typo3/cms-core/main/en-us/Changelog/12.4/Deprecation-100173-VariousMethodsAndPropertiesInUserAuthenticationClassesNowInternal.html

Resolves: https://gitlab.com/jigal/t3adminer/-/issues/32